fix: Adjusted handler dashboard fetchs
